Provide onViewClicked for member in CoreFamilyProfileMemberFragment so that it is avaiable in chw/hf fragments
