+++ opensrp-chw/src/main/java/org/smartregister/chw/fragment/FamilyProfileMemberFragment.java
@@ -2,7 +2,6 @@ package org.smartregister.chw.fragment;
 
 import android.os.Bundle;
 
-import org.smartregister.chw.activity.FamilyProfileActivity;
 import org.smartregister.chw.core.fragment.CoreFamilyProfileMemberFragment;
 import org.smartregister.chw.model.FamilyProfileMemberModel;
 import org.smartregister.chw.provider.ChwMemberRegisterProvider;
@@ -40,23 +39,5 @@ public class FamilyProfileMemberFragment extends CoreFamilyProfileMemberFragment
         String familyHead = getArguments().getString(Constants.INTENT_KEY.FAMILY_HEAD);
         String primaryCareGiver = getArguments().getString(Constants.INTENT_KEY.PRIMARY_CAREGIVER);
         presenter = new BaseFamilyProfileMemberPresenter(this, new FamilyProfileMemberModel(), null, familyBaseEntityId, familyHead, primaryCareGiver);
-    }
-
-    @Override
-    protected void onViewClicked(android.view.View view) {
-        super.onViewClicked(view);
-        int i = view.getId();
-        if (i == org.smartregister.chw.core.R.id.patient_column) {
-            if (view.getTag() != null && view.getTag(org.smartregister.family.R.id.VIEW_ID) == CLICK_VIEW_NORMAL) {
-                getFamilyProfileActivity().goToProfileActivity(view, getArguments());
-            }
-        } else if (i == org.smartregister.chw.core.R.id.next_arrow && view.getTag() != null &&
-                view.getTag(org.smartregister.family.R.id.VIEW_ID) == CLICK_VIEW_NEXT_ARROW) {
-            getFamilyProfileActivity().goToProfileActivity(view, getArguments());
-        }
-    }
-
-    private FamilyProfileActivity getFamilyProfileActivity() {
-        return (FamilyProfileActivity) getActivity();
     }
 }
